Move memex-specific storex extensions out of init logic
- moved out of `search/storex` to `search/memex-storex` module
- test env will need to extend this differently (deleteDB is not on window, for example)
